+++ test/sync/src/androidTest/kotlin/io/realm/kotlin/test/mongodb/shared/internal/KtorNetworkTransportTest.kt
@@ -24,7 +24,7 @@ import io.realm.kotlin.internal.util.use
 import io.realm.kotlin.mongodb.internal.KtorNetworkTransport
 import io.realm.kotlin.test.mongodb.TEST_SERVER_BASE_URL
 import io.realm.kotlin.test.mongodb.util.BaasClient
-import io.realm.kotlin.test.mongodb.util.KtorAppTester.initialize
+import io.realm.kotlin.test.mongodb.util.KtorTestAppInitializer.initialize
 import kotlinx.coroutines.channels.Channel
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
